agent/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0

# Pydantic for configuration and data validation
pydantic==2.9.2
pydantic-settings==2.5.2

# LINE Bot SDK
line-bot-sdk==3.13.0

# AWS SDK for Python (Boto3) - for AgentCore Gateway integration
boto3==1.35.22

# HTTP Client
httpx==0.27.2
